GCS: add support for tracker messages and cmds:
this commit affects GCS.h and ap_message.h from libraries, and
gcs_mavlink from copter. We aded here a new group for frequency
rate params, TRACKER, and support for managing the rate of this one
individually on a separate group. also handlers in gcs_mavlink for
the commands comming from the gcs